pyproject.toml
[tool.poetry]
name = "coursework"
version = "0.1.0"
description = ""
authors = ["FlakR <neykerez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.1.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
mypy = "^1.10.1"
isort = "^5.13.2"
black = "^24.4.2"
flake8 = "^7.1.0"
yfinance = "^0.2.40"
xlrd = "^2.0.1"

[tool.poetry.group.lint.dependencies]
mypy = "^1.9.0"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
ignore_missing_imports = true
exclude = 'venv'


[tool.isort]
line_length = 120
profile = "black"
multi_line_output = 3


[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
)
)
'''



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
